Split up `DatasetActionButton`
