+++ web/src/features/datasets/components/DatasetsTable.tsx
@@ -1,7 +1,8 @@
 import { DataTable } from "@/src/components/table/data-table";
 import TableLink from "@/src/components/table/table-link";
 import { type LangfuseColumnDef } from "@/src/components/table/types";
-import { DatasetActionButton } from "@/src/features/datasets/components/DatasetActionButton";
+import { DeleteDatasetButton } from "@/src/features/datasets/components/DeleteDatasetButton";
+import { UpdateDatasetButton } from "@/src/features/datasets/components/UpdateDatasetButton";
 import { DatasetSchemaHoverCard } from "@/src/features/datasets/components/DatasetSchemaHoverCard";
 import { useDetailPageLists } from "@/src/features/navigate-detail-pages/context";
 import { api } from "@/src/utils/api";
@@ -454,8 +455,7 @@ export function DatasetsTable(props: { projectId: string }) {
 
         return (
           <div className="flex items-center gap-1">
-            <DatasetActionButton
-              mode="update"
+            <UpdateDatasetButton
               projectId={props.projectId}
               datasetId={key.id}
               datasetName={row.original.folderPath}
@@ -469,8 +469,7 @@ export function DatasetsTable(props: { projectId: string }) {
               variant="ghost"
               size="icon-xs"
             />
-            <DatasetActionButton
-              mode="delete"
+            <DeleteDatasetButton
               projectId={props.projectId}
               datasetId={key.id}
               datasetName={row.original.folderPath}
